feat: Add file creation functionality and update data types
Added the ability to create a new file in the partition with error handling. Updated data types for better memory usage efficiency. Included placeholders for read, write, and delete file functions in the header file.

+++ src/ufs.priv.h
@@ -32,8 +32,8 @@ typedef struct {
 typedef struct {
     int fd;
     super_bloc_t super_bloc;
-    uint32_t *data_bitmap;
-    uint32_t *inode_bitmap;
+    uint8_t *data_bitmap;
+    uint8_t *inode_bitmap;
     opened_file_t opened_files[MAX_OPENED_FILES];
     uint16_t nb_opened_files;
 } partition_t;
